refactor: convert browser builtin into proper package (models, tools, facade)
